+++ config/cache.php
@@ -1,19 +1,19 @@
 <?php
 /**
- * This file is part of MeYoutube.
+ * This file is part of me-cms-youtube.
  *
- * MeYoutube is free software: you can redistribute it and/or modify
+ * me-cms-youtube is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
- * MeYoutube is distributed in the hope that it will be useful,
+ * me-cms-youtube is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Affero General Public License for more details.
  *
  * You should have received a copy of the GNU Affero General Public License
- * along with MeYoutube.  If not, see <http://www.gnu.org/licenses/>.
+ * along with me-cms-youtube.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
  * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
@@ -21,19 +21,16 @@
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  */
 
-if (!defined('ME_YOUTUBE_CACHE')) {
-    define('ME_YOUTUBE_CACHE', CACHE . 'me_youtube' . DS);
+if (!defined('ME_CMS_YOUTUBE_CACHE')) {
+    define('ME_CMS_YOUTUBE_CACHE', CACHE . 'me_cms_youtube' . DS);
 }
 
-//Default options (with File engine)
-$options = [
-    'className' => 'File',
-    'duration' => '+999 days',
-    'path' => ME_YOUTUBE_CACHE,
-    'prefix' => '',
-    'mask' => 0777,
-];
-
 return ['Cache' => [
-    'videos' => am($options, ['path' => ME_YOUTUBE_CACHE . 'videos'])
+    'videos' => [
+        'className' => 'File',
+        'duration' => '+999 days',
+        'path' => ME_CMS_YOUTUBE_CACHE . 'videos',
+        'prefix' => '',
+        'mask' => 0777,
+    ],
 ]];