+++ src/Model/Table/VideosCategoriesTable.php
@@ -1,32 +1,30 @@
 <?php
 /**
- * This file is part of MeYoutube.
+ * This file is part of me-cms-youtube.
  *
- * MeYoutube is free software: you can redistribute it and/or modify
+ * me-cms-youtube is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
- * MeYoutube is distributed in the hope that it will be useful,
+ * me-cms-youtube is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Affero General Public License for more details.
  *
  * You should have received a copy of the GNU Affero General Public License
- * along with MeYoutube.  If not, see <http://www.gnu.org/licenses/>.
+ * along with me-cms-youtube.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
  * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  */
-namespace MeYoutube\Model\Table;
+namespace MeCmsYoutube\Model\Table;
 
 use Cake\ORM\Query;
 use Cake\ORM\RulesChecker;
-use Cake\Validation\Validator;
 use MeCms\Model\Table\AppTable;
-use MeYoutube\Model\Entity\VideosCategory;
 
 /**
  * VideosCategories model
@@ -106,15 +104,15 @@ class VideosCategoriesTable extends AppTable
         $this->primaryKey('id');
 
         $this->belongsTo('Parents', [
-            'className' => 'MeYoutube.VideosCategories',
+            'className' => 'MeCmsYoutube.VideosCategories',
             'foreignKey' => 'parent_id',
         ]);
         $this->hasMany('Childs', [
-            'className' => 'MeYoutube.VideosCategories',
+            'className' => 'MeCmsYoutube.VideosCategories',
             'foreignKey' => 'parent_id',
         ]);
         $this->hasMany('Videos', [
-            'className' => 'MeYoutube.Videos',
+            'className' => 'MeCmsYoutube.Videos',
             'foreignKey' => 'category_id',
         ]);
 
@@ -125,10 +123,10 @@ class VideosCategoriesTable extends AppTable
     /**
      * Default validation rules
      * @param \Cake\Validation\Validator $validator Validator instance
-     * @return \MeYoutube\Model\Validation\VideosCategoryValidator
+     * @return \MeCmsYoutube\Model\Validation\VideosCategoryValidator
      */
     public function validationDefault(\Cake\Validation\Validator $validator)
     {
-        return new \MeYoutube\Model\Validation\VideosCategoryValidator;
+        return new \MeCmsYoutube\Model\Validation\VideosCategoryValidator;
     }
 }