+++ src/View/Helper/MenuHelper.php
@@ -1,19 +1,19 @@
 <?php
 /**
- * This file is part of MeYoutube.
+ * This file is part of me-cms-youtube.
  *
- * MeYoutube is free software: you can redistribute it and/or modify
+ * me-cms-youtube is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
- * MeYoutube is distributed in the hope that it will be useful,
+ * me-cms-youtube is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Affero General Public License for more details.
  *
  * You should have received a copy of the GNU Affero General Public License
- * along with MeYoutube.  If not, see <http://www.gnu.org/licenses/>.
+ * along with me-cms-youtube.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
  * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
@@ -21,7 +21,7 @@
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  * @see         MeCms\View\Helper\MenuBuilderHelper
  */
-namespace MeYoutube\View\Helper;
+namespace MeCmsYoutube\View\Helper;
 
 use Cake\View\Helper;
 
@@ -52,15 +52,15 @@ class MenuHelper extends Helper
     public function videos()
     {
         $menu = [
-            $this->Html->link(__d('me_youtube', 'List videos'), [
+            $this->Html->link(__d('me_cms_youtube', 'List videos'), [
                 'controller' => 'Videos',
                 'action' => 'index',
-                'plugin' => MEYOUTUBE,
+                'plugin' => ME_CMS_YOUTUBE,
             ]),
-            $this->Html->link(__d('me_youtube', 'Add video'), [
+            $this->Html->link(__d('me_cms_youtube', 'Add video'), [
                 'controller' => 'Videos',
                 'action' => 'add',
-                'plugin' => MEYOUTUBE,
+                'plugin' => ME_CMS_YOUTUBE,
             ]),
         ];
 
@@ -69,15 +69,15 @@ class MenuHelper extends Helper
             $menu[] = $this->Html->link(__d('me_cms', 'List categories'), [
                 'controller' => 'VideosCategories',
                 'action' => 'index',
-                'plugin' => MEYOUTUBE,
+                'plugin' => ME_CMS_YOUTUBE,
             ]);
             $menu[] = $this->Html->link(__d('me_cms', 'Add category'), [
                 'controller' => 'VideosCategories',
                 'action' => 'add',
-                'plugin' => MEYOUTUBE,
+                'plugin' => ME_CMS_YOUTUBE,
             ]);
         }
 
-        return [$menu, __d('me_youtube', 'Videos'), ['icon' => 'film']];
+        return [$menu, __d('me_cms_youtube', 'Videos'), ['icon' => 'film']];
     }
 }